Revert "[test]: added gesture test"
