--- hand_gesture.js
@@ -1,123 +1,0 @@
-var start_tracking = false;
-var x1 = 0;
-var x2 = 0;
-
-function onResults(results) {
-    if (results.multiHandLandmarks.length != 0) {
-        var [lsit, box] = findhandpos(results.multiHandLandmarks[0]);
-    }
-    else {
-        var lsit = [];
-        if (start_tracking == true) {
-            //Stop Tracking
-            start_tracking = false;
-            if (x1[0] > x2[0]) {
-                //Gesture 1
-                return "right swipe"
-            }
-            else {
-                //Gesture 2
-                return "left swipe"
-            }
-        }
-    }
-
-    if (lsit.length != 0) {
-
-        var fings = detect_fingersup(lsit);
-
-        // console.log(fings);
-
-        if (fings[1] == true && fings[3] == false && fings[4] == false) {
-            if (start_tracking == false) {
-                //start tracking
-                start_tracking = true;
-                x1 = lsit[8].slice(1, 2);
-                //console.log(x1)
-            }
-            else {
-                x2 = lsit[8].slice(1, 2);
-            }
-        }
-        else {
-
-            start_tracking = false;
-
-            if (fings[0] == true && fings[1] == true && fings[2] == true && fings[3] == true && fings[4] == true) {
-                //Gesture 3
-                return "all five fingers"
-
-            }
-
-            else if (fings[0] == true && fings[1] == false && fings[2] == false && fings[3] == false && fings[4] == false) {
-                var y1 = lsit[4].slice(2, 3);
-                var y2 = lsit[2].slice(2, 3);
-                if (y1[0] > y2[0]) {
-                    //Gesture 4
-                    return "Thumbs Down"
-                }
-                else {
-                    //Gesture 5
-                    return "Thumbs Up"
-                }
-            }
-        }
-    }
-
-    return "Not Detected"
-}
-
-function findhandpos(landmarks) {
-    var xlist = [];
-    var ylist = [];
-    var lmlist = [];
-    var bbox = [];
-    for (const [index, element] of landmarks.entries()) {
-        var [h, w, c] = [900, 1600, 3];
-        var cx = parseInt(element.x * w);
-        var cy = parseInt(element.y * h);
-        xlist.push(cx);
-        ylist.push(cy);
-        lmlist.push([index, cx, cy]);
-    }
-    var xmin = Math.min(xlist);
-    var xmax = Math.max(xlist);
-    var ymin = Math.min(ylist);
-    var ymax = Math.max(ylist);
-    bbox = [xmin, ymin, xmax, xmin];
-
-    return [lmlist, bbox];
-}
-
-function detect_fingersup(lmlist) {
-    var fingers = [];
-    var ids = [4, 8, 12, 16, 20];
-    if (lmlist[ids[0]][1] < lmlist[ids[0] - 2][1]) {
-        fingers.push(1);
-    }
-    else {
-        fingers.push(0);
-    }
-    for (let i = 1; i < 5; i++) {
-        if (lmlist[ids[i]][2] < lmlist[ids[i] - 2][2]) {
-            fingers.push(1);
-        }
-        else {
-            fingers.push(0);
-        }
-    }
-    return fingers;
-}
-
-
-try {
-    module.exports = {onResults, findhandpos, detect_fingersup}
-} catch (error) {}
-
-// //export onResults = onResults
-// try {
-//     //
-   
-// } catch (error) {
-
-// }
